Code structure change - Convert util functions to class methods. Change file name to restrict the context. - Split analysis file
